pyproject.toml
[project]
name = "video-insight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.11"
dependencies = [
    "funasr>=1.0.0",
    "modelscope>=1.10.0",
    "requests>=2.32.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "tqdm>=4.67.0",
    "lark-oapi>=1.5.2",
    "opencv-python>=4.13.0.90",
    "imageio-ffmpeg>=0.4.9",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
]

[tool.uv]
index = [
    { name = "pytorch-cpu", url = "https://download.pytorch.org/whl/cpu", explicit = true },
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchaudio = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[project.scripts]
vt = "runners.cli:main"
vt-bot = "runners.bot_runner:main"
vt-server = "runners.server_runner:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

